Correcting output wavelengths
The L6 wavelength has been updated in line 254. The output wavelength for filter L6 was previously 440 nm, but according to Bell et al (2020), the correct wavelength is 442 nm and the code has been updated accordingly.

This does not affect the convolution calculation because no changes have been made to the filter band passes. This is only a change to the output wavelength.